Features & Tools    * Full Metadata Retrieval: Updated list_tasks and list_project_view_tasks to return complete metadata      (descriptions, assignees, buckets) by default, reducing multi-turn overhead.    * Project View Support: Added get_project and list_project_view_tasks tools to enable efficient bulk data      gathering via specific project views (List, Kanban, etc.).    * Token Optimization: Implemented an include_descriptions toggle (default True) for task listing tools,      allowing agents to opt-out for bulk administrative tasks.    * MCP UI Refactor: Optimized all 21 tools with single-line docstrings and structured Annotated parameters      for cleaner hermes output and reduced token usage.
  Bug Fixes
   * Resilient Expansion: Fixed 412 Precondition Failed errors in get_task by implementing an automated
     fallback to manual sub-resource retrieval (attachments, reminders, assignees) if bulk expansion is
     rejected.
   * Type Mismatch: Updated position, index, and percent_done field types to float to support the fractional
     values returned by Vikunja for Kanban ordering.
   * Environment Stability: Resolved MCP server startup crashes by correcting missing typing and pydantic
     imports.

  Maintenance & Reliability
   * Validation Workflow: Formalized a mandatory "Clean-Room Subagent Validation" requirement in .ai/rules.md
     and created a dedicated subskill guide in .ai/mcp_validation.md.
   * Test Coverage: Added and verified integration tests for task hierarchy, expansion fallbacks, and
     description toggling.

+++ vikunja_python/core/models/task.py
@@ -42,13 +42,13 @@ class Task(VikunjaBaseModel):
     title: str = Field(..., max_length=255, description="Task title (required)")
     description: Optional[str] = Field(None, max_length=10000, description="Task description (markdown)")
     identifier: str = Field(..., description="Task identifier within project (e.g., '#1')")
-    index: int = Field(0, description="Task index for ordering within view/bucket")
+    index: float = Field(0.0, description="Task index for ordering within view/bucket")
     project_id: int = Field(..., description="Parent project ID")
     
     # Status Fields (3 fields)
     done: bool = Field(False, description="Completion status")
     done_at: Optional[datetime] = Field(None, description="Timestamp when task was marked done")
-    percent_done: int = Field(0, ge=0, le=100, description="Completion percentage (0-100)")
+    percent_done: float = Field(0.0, ge=0.0, le=100.0, description="Completion percentage (0-100)")
     
     # Priority & Importance (3 fields)
     priority: int = Field(0, ge=0, le=5, description="Task priority (0-5, higher = more important)")
@@ -70,7 +70,7 @@ class Task(VikunjaBaseModel):
     
     # Positioning Fields (3 fields) - For Kanban/board views
     bucket_id: int = Field(0, description="Bucket ID in board view")
-    position: int = Field(0, description="Position within bucket")
+    position: float = Field(0.0, description="Position within bucket")
     
     # Nested Objects (4 fields)
     assignees: Optional[list[User]] = Field(default_factory=list, description="Users assigned to this task")
@@ -187,7 +187,7 @@ class TaskUpdateRequest(VikunjaBaseModel):
     due_date: Optional[datetime] = None
     start_date: Optional[datetime] = None
     end_date: Optional[datetime] = None
-    percent_done: Optional[int] = Field(None, ge=0, le=100)
+    percent_done: Optional[float] = Field(None, ge=0.0, le=100.0)
     is_favorite: Optional[bool] = None
     hex_color: Optional[str] = None
     
@@ -197,7 +197,7 @@ class TaskUpdateRequest(VikunjaBaseModel):
     
     # Positioning
     bucket_id: Optional[int] = None
-    position: Optional[int] = None
+    position: Optional[float] = None
     
     def model_dump_for_api(self) -> dict:
         """Convert to dict, excluding None values."""
